Split up query tests: - In response to feedback from @Geeber in a previous PR. - KeenQueryTest was getting unwieldy. There's still plenty more to do to clean up existing tests, but at least get these sharing a common set of base functionality before adding new tests. - In preparation for adding tests for Saved/Cached Queries, Access Keys and Datasets.
